[#3840] Tests: extract shared account seeding into AccountSeedHelper
seed_account_with_password had four copies (MfaFlowHelper plus three full/
specs) and seed_existing_account three, all behaviourally identical. Move both
to support/account_seed_helper.rb, auto-included into type: :integration
alongside AuthRequestHelper and included explicitly by MfaFlowHelper so the
full_mfa lane stands on its own.

The why-notes scattered across the copies (Verified status satisfying
_account_from_login and open_account?, external_id pairing the accounts row to
its Customer for flows that probe customer state, Argon2 costs mirroring
config/features/argon2.rb, the password as OTP-provisioning vehicle) are
consolidated in the new helper's header.
