+++ apps/web/auth/spec/integration/full/identities_routes_spec.rb
@@ -63,20 +63,7 @@ RSpec.describe 'Linked identities management API (#3840 Phase 2)', type: :integr
   # Helpers
   # ==========================================================================
 
-  def seed_account_with_password(email, password: AuthTestConstants::TEST_PASSWORD)
-    normalized = OT::Utils.normalize_email(email)
-    customer   = Onetime::Customer.new(email: normalized)
-    customer.save
-    account_id = auth_db[:accounts].insert(
-      email: normalized,
-      status_id: AuthTestConstants::STATUS_VERIFIED,
-      external_id: customer.extid,
-    )
-    require 'argon2'
-    hasher     = Argon2::Password.new(t_cost: 1, m_cost: 5, p_cost: 1)
-    auth_db[:account_password_hashes].insert(id: account_id, password_hash: hasher.create(password))
-    account_id
-  end
+  # seed_account_with_password comes from support/account_seed_helper.rb.
 
   def remove_password(account_id)
     auth_db[:account_password_hashes].where(id: account_id).delete
