+++ apps/web/auth/spec/integration/full/omniauth_connect_link_spec.rb
@@ -104,28 +104,9 @@ RSpec.describe 'OmniAuth authenticated identity connect (#3840 Phase 2)', type: 
     allow(Onetime.auth_config).to receive(:allow_platform_fallback_for_tenants?).and_return(true)
   end
 
-  # Seed a VERIFIED account WITHOUT a password (SSO-only / victim account).
-  def seed_existing_account(email)
-    normalized = OT::Utils.normalize_email(email)
-    customer   = Onetime::Customer.new(email: normalized)
-    customer.save
-    auth_db[:accounts].insert(
-      email: normalized,
-      status_id: AuthTestConstants::STATUS_VERIFIED,
-      external_id: customer.extid,
-    )
-  end
-
-  # Seed a VERIFIED account WITH an Argon2 password hash so csrf_login can
-  # establish a real authenticated session for it. Cost params match the test
-  # config in config/features/argon2.rb.
-  def seed_account_with_password(email, password: AuthTestConstants::TEST_PASSWORD)
-    account_id = seed_existing_account(email)
-    require 'argon2'
-    hasher     = Argon2::Password.new(t_cost: 1, m_cost: 5, p_cost: 1)
-    auth_db[:account_password_hashes].insert(id: account_id, password_hash: hasher.create(password))
-    account_id
-  end
+  # seed_existing_account (the SSO-only / victim account) and
+  # seed_account_with_password (the subject csrf_login can authenticate as)
+  # come from support/account_seed_helper.rb.
 
   # Establish a session, fetch the CSRF token, then POST a JSON login. The full
   # Rack app enforces CSRF, so the shrimp token is required.
